New function: Change the default audio stream which's volume is changed when using the volume rocker.
The standard default audio stream is STREAM_RING. When the user uses the volume rocker and no special stream is active (no app requested a audio stream change, e.g. no audio is playing) the ringer volume is changed.
So the ringer volume is often changed accidentally. This change gives the user the option to change the default stream between STREAM_RING and STREAM_MUSIC in Audio Settings.